requirements.txt
absl-py==2.1.0
accelerate==1.7.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==4.2.2
anyio==4.6.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==24.2.0
audioread==3.0.1
babel==2.16.0
bcrypt==4.3.0
beautifulsoup4==4.12.3
bert-score==0.3.13
bleach==6.1.0
blinker==1.8.2
cachetools==5.5.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
configparser==7.2.0
contourpy==1.3.1
cycler==0.12.1
datasets==3.6.0
debugpy==1.8.6
decorator==5.1.1
deep-translator==1.11.4
defusedxml==0.7.1
dill==0.3.8
einops==0.8.1
entrypoints==0.4
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.13.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
flatbuffers==25.1.24
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.6.0
fsspec==2024.6.1
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.70.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.31.2
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
# Editable install with no version control (inittest==0.1.0)
-e c:\users\tahmi\documents\work\learn\inittest
ipykernel==6.29.5
ipython==8.28.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kaggle==1.7.4.5
keras==3.8.0
kiwisolver==1.4.8
kymatio==0.3.0
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
llvmlite==0.44.0
lxml==5.4.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
multiprocess==0.70.16
namex==0.0.8
narwhals==1.39.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
notebook==7.2.2
notebook_shim==0.2.4
numba==0.61.2
numpy==1.26.4
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.0
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
Pillow==9.5.0
platformdirs==4.3.6
pooch==1.8.2
portalocker==3.1.1
progressbar==2.5
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.3.1
protobuf==4.25.8
psutil==6.0.0
pure_eval==0.2.3
pyarrow==20.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.10.1
Pympler==1.1
pyparsing==3.2.1
pyrefly==0.16.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-slugify==8.0.4
pytz==2024.2
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rouge_score==0.1.2
rpds-py==0.20.0
sacrebleu==2.5.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==4.1.0
sentencepiece==0.2.0
setuptools==80.9.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.2
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
spectral==0.24
SpeechRecognition==3.13.0
SQLAlchemy==2.0.39
stack-data==0.6.3
streamlit==1.22.0
streamlit-audiorec==0.1.3
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-hub==0.16.1
tensorflow_intel==2.18.0
termcolor==2.5.0
terminado==0.18.1
# Editable install with no version control (test==0.1.0)
-e c:\users\tahmi\documents\work\learn\inittest
# Editable install with no version control (test1==0.1.0)
-e c:\users\tahmi\documents\work\learn\inittest
text-unidecode==1.3
tf_keras==2.19.0
threadpoolctl==3.5.0
timm==1.0.15
tinycss2==1.3.0
tokenizers==0.21.1
toml==0.10.2
toolz==1.0.0
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchvision==0.21.0+cu118
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.2
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.3.1
umap==0.1.1
unicode==2.9
Unidecode==1.3.8
uri-template==1.3.0
urllib3==2.2.3
validators==0.35.0
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.4
wheel==0.45.1
wordcloud==1.9.4
wordninja==2.0.0
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
